Move ldap logic into ldap authentication provider
